src/services/gads_reporting/cloud_functions/ads_report/requirements.txt
Flask==2.1.3
google-ads==18.0.0
jsonschema==4.9.1
pandas==1.4.3
pandas-gbq==0.17.7
